analytics: batch inserts to increase performance by some 300 %
The improvement here is due to a number of factors:
- If we re-construct the statement each time, as we did before,
  we re-compile the query *every single time* from sqla DSL
  to SQL
- By issuing each query alone, we block for each query until the
  database has processed it
- The database has to parse each query independently and generate
  a query plan for it each time (I don't think postgres does
  automatic plan caching)

So by batching these updates we remove all of these:
- We generate SQL a few times only (once per batch)
- The DB parses and plans one query per ~1000 inserts
- We do ~one round trip per ~1000 inserts

There is another free ~15 % performance improvement by making
the ProgressBar smarter, but I think that might get replaced
by a redis pub/sub anyway to be able to monitor progress
from the web frontend. Or not. Who knows.