dal.daily_stats: join instead of correlated query
This does mean that the group-wise sorting and limiting is done in
the application, which should not be a biggy. It *can* be done in
SQL as well, but basically any "group-wise X" operation is a bit
complicated.

For the time being and the current data set, this is more than
an order of magnitude faster than the correlated query, which
in turn was some five or so times faster than an application-correlated
query.