analytics: avoid slow "[unknown subset] JOIN [big table]" query
The issue here is basically that the database can't really
know how exactly the PID foreign key relates (we have a 1:1 constraint),
so it *has to* join against the big post_contents table in full.

By applying the same predicate to both tables only a small amount of
rows need to be considered for the join.