crawler: check if TNU entry exists
this can happen in a number of ways (running twice w/o processing TNU),
but in this case it happened because TID#99378 was moved from BID#95,
where it was first seen and added to TNU in --board-id=all mode,
to BID#69 where it was seen the second time. Due to --board-id=all,
no TNU processing happened between those sightings.