analytics: fix potential deadlock
Previously a simple "session.close()" was observed to cause a hang
in psycopg/libpq via rollback() with a trace like this:

Current thread 0x00007f4dbeef7540 (most recent call first):
  File "sqlalchemy/engine/default.py", line 538 in do_rollback
  File "sqlalchemy/engine/base.py", line 733 in _rollback_impl
  File "sqlalchemy/engine/base.py", line 1754 in _do_rollback
  File "sqlalchemy/engine/base.py", line 1716 in rollback
  File "sqlalchemy/pool/base.py", line 878 in _reset
  File "sqlalchemy/pool/base.py", line 693 in _finalize_fairy
  File "sqlalchemy/pool/base.py", line 860 in _checkin
  File "sqlalchemy/pool/base.py", line 1012 in close
  File "sqlalchemy/engine/base.py", line 896 in close
  File "sqlalchemy/orm/session.py", line 588 in close
  File "sqlalchemy/orm/session.py", line 1337 in _close_impl
  File "sqlalchemy/orm/session.py", line 1298 in close
  File "potstats2/analytics.py", line 263 in analyze_posts
  File "potstats2/analytics.py", line 40 in main
  File "click/core.py", line 555 in invoke

Meanwhile invalidate() in conjunction with an Engine.dispose call
should be safe, since invalidate() is meant to be used when
the database connection may not be in a usable state any more.
close() in contrast may attempt cleanup, like issuing a ROLLBACK
as shown here.